+++ examples/resources/forgejo_repository_actions_variable/resource.tf
@@ -1,0 +1,6 @@
+resource "forgejo_repository_actions_variable" {
+  data       = "value"
+  name       = "test"
+  owner      = "adyxax"
+  repository = "example"
+}