+++ CHANGELOG.md
@@ -10,5 +10,6 @@ All notable changes to this project will be documented in this file.
 - Added organizations data-source.
 - Added repositories data-source.
 - Added repository actions secret resource.
+- Added repository actions variable resource.
 - Added teams data-source.
 - Added users data-source.